requirements.txt
openai>=1.30.0
python-dotenv>=1.0.1
typer[all]>=0.12.3
rich>=13.7.1
pydantic>=2.7.1
